Pre Final - Impl Noficiacao Complete

+++ src/br/com/pluri/eventor/dao/AtividadeDAO.java
@@ -1,6 +1,5 @@
 package br.com.pluri.eventor.dao;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.Query;
@@ -34,8 +33,8 @@ public interface AtividadeDAO extends BaseDAO<Atividade> {
 	public List<Atividade> findAllAtividadeByUsuario (@Param("id") Long idUsu);
 	
 	@Query(value = "SELECT COUNT(*) FROM TBL_USUARIO_ATIVIDADE WHERE STATUS IN ('Pendente', 'Aprovado') AND ID_ATIVI = :idAtiv", nativeQuery = true)
-	public BigInteger qtdInscritoInAtividade (@Param("idAtiv") Long idAtiv);
+	public int qtdInscritoInAtividade (@Param("idAtiv") Long idAtiv);
 	
 	@Query(value = "SELECT COUNT(*) FROM TBL_USUARIO_ATIVIDADE WHERE STATUS = :status AND ID_ATIVI = :idAtiv", nativeQuery = true)
-	public BigInteger qtdInscritoSitInAtividade (@Param("status") String status, @Param("idAtiv") Long idAtiv);
+	public int qtdInscritoSitInAtividade (@Param("status") String status, @Param("idAtiv") Long idAtiv);
 }
